requirements.txt
pandas
torch
black
attrs
sacrebleu
nltk==3.7
expiringdict
attrs
fastapi
pretty_html_table
pytest
black
isort
autoflake
pre-commit
uvicorn
